+++ test/test_nose/test_nose_example.py
@@ -2,11 +2,13 @@
 
 from __future__ import unicode_literals
 
-from unittest import TestCase, expectedFailure, skip
+from unittest import TestCase, skip
+
+from genty import genty, genty_dataset
+from nose.tools import raises
 
 from flaky import flaky
 
-from genty import genty, genty_dataset
 
 # This is an end-to-end example of the flaky package in action. Consider it
 # a live tutorial, showing the various features in action.
@@ -17,9 +19,8 @@ class ExampleTests(TestCase):
 
     def test_non_flaky_thing(self):
         """Flaky will not interact with this test"""
-        pass
 
-    @expectedFailure
+    @raises(AssertionError)
     def test_non_flaky_failing_thing(self):
         """Flaky will also not interact with this test"""
         self.assertEqual(0, 1)
@@ -27,8 +28,7 @@ class ExampleTests(TestCase):
     @flaky(3, 2)
     def test_flaky_thing_that_fails_then_succeeds(self):
         """
-        Flaky will run this test 3 times.
-        It will fail once and then succeed twice.
+        Flaky will run this test 3 times. It will fail once and then succeed twice.
         """
         self._threshold += 1
         if self._threshold < 1:
@@ -39,8 +39,7 @@ class ExampleTests(TestCase):
     @flaky(3, 2)
     def test_flaky_thing_that_succeeds_then_fails_then_succeeds(self):
         """
-        Flaky will run this test 3 times.
-        It will succeed once, fail once, and then succeed one more time.
+        Flaky will run this test 3 times. It will succeed once, fail once, and then succeed one more time.
         """
         self._threshold += 1
         if self._threshold == 1:
@@ -49,7 +48,6 @@ class ExampleTests(TestCase):
     @flaky(2, 2)
     def test_flaky_thing_that_always_passes(self):
         """Flaky will run this test twice.  Both will succeed."""
-        pass
 
     @skip("This really fails! Remove this decorator to see the test failure.")
     @flaky()
@@ -64,8 +62,7 @@ class ExampleFlakyTests(TestCase):
 
     def test_flaky_thing_that_fails_then_succeeds(self):
         """
-        Flaky will run this test twice.
-        It will fail once and then succeed.
+        Flaky will run this test twice. It will fail once and then succeed.
         """
         self._threshold += 1
         if self._threshold < 1:
@@ -79,7 +76,6 @@ def test_function():
     Nose will import this function and wrap it in a :class:`FunctionTestCase`.
     It's included in the example to make sure flaky handles it correctly.
     """
-    pass
 
 
 @flaky